LayoutTests/fast/js/regexp-literals-are-constants-expected.txt

This test ensures that regeular expression literals are constants, and so persist over multiple executions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS currentRegExp is lastRegExp
PASS currentRegExp is lastRegExp
PASS returnRegExpLiteral() is returnRegExpLiteral()
PASS returnConditionalRegExpLiteral(true) is returnConditionalRegExpLiteral(true)
PASS returnConditionalRegExpLiteral(false) is returnConditionalRegExpLiteral(false)
PASS returnConditionalRegExpLiteral(true) === returnConditionalRegExpLiteral(false) is false
PASS returnRegExpLiteral().someAddedProperty is true
PASS successfullyParsed is true

TEST COMPLETE
